refactor(entranceDoc): refactor documents
